@@ -2,7 +2,7 @@ class SamlController < ApplicationController
 
   def consume
     response = OneLogin::RubySaml::Response.new(params[:SAMLResponse])
-    render :text => response.name_id
+    render plain: response.name_id
   end
 
 end
