@@ -30,8 +30,11 @@ RailsApp::Application.configure do
   #config.active_record.auto_explain_threshold_in_seconds = 0.5
 
   # Do not compress assets
-  config.assets.compress = false
+  # config.assets.compress = false
 
   # Expands the lines which load the assets
-  config.assets.debug = true
+  # config.assets.debug = true
+
+  config.eager_load = false
+  config.hosts << 'foo.example.com'
 end
